increment logical filter before filtering
This is needed because when Salesforce throws
an error if there is a column not included
in the logical filter, i.e., we said to it
to not include values that we have already seen,
but without telling how this condition relates to
the others.

Now, we add a "AND 5" if the filter has 4 columns,
like "1 AND 2 AND 3 AND 4", everytime we try to
surpass the 2000 row limit.