fix wrong order of cells in matrix reports
The old way worked for a very small subset of cells. For instance, it
would sort an array with ['0_0!0_1', '0_0!0_2', '0_0!10_0'] as
['0_0!0_1', '0_0!10_0', '0_0!0_2'], which is not what I want.

This was solved by passing as the key value the
distutils.version.LooseVersion class, which sorts an array with
version-like strings, just what we wanted.

Also, I filtered the factMap beforehand because an error was thrown
otherwise (something like TypeError: '<' not supported between instances
of 'str' and 'int'), probably because there was a 'T' in the string.
Once I filtered out the grandTotals and subTotals, it worked.