improve test coverage
This test the case in which a currency data type
does not come with a "value" attribute, instead of
it being separated in "amount" and "currency", i.e.

We were testing only this case:

{
    "label": "$16,000.01",
    "value": {
        "amount": 16000.01,
        "currency": null
    }
}

But there might be also cases like this, who knows:

{
    "label": "$16,000.01",
    "value": 16000.01
}

Also, test getting a matrix column labels, as this wasn't tested
by the other tests.