fix filters argument conflicting with filters module
Fix mistake of naming the filters helper module the same
as the filters argument in Reportforce.get_report.

So when I called filters.set_report it would actually do
[].set_report by default since filters in that function was
set to an empty list.

I then renamed the filters module to sf_filters, much uglier.
To avoid this ugliness, I preferred to import only the relevant
used by a module.