improve response error handling
Do nothing in case of a KeyError, which means that
there is no error message or error code in the response
body.

Also ignore error if it was a ValueError, probably due
to a malformed JSON string (such as binary data).

Unfortunately, it would be ugly to try to be specific
about this because there is not a universal JSONDecodeError
across libraries. But AFAIK every one of them is an instance
of a ValueError.