fix StopIteration breaking test suite
To simulate the process of getting a report
until all data has been returned, we have to
generate a first report with its metadata value
'allData' set to False and then becoming True.

I did this with the side_effect attribute of a
mocked Reportforce.session.post.

But this broke my test suite in Python >= 3.7,
because, apparently, the approach I was using
raised an StopIteration to tell there is no more
reports to generate, and this translates to a
RuntimeError in later versions of Python.

So, now what I did was to create the function
generate_reports in the utils/mocks folder, that
gets a report, copies it, set the allData
attribute to False, and then generate the first report
as is, i.e., with allData set to True.

No more StopIteration in this way.