Add parsers.map_columns_to_dtypes
This simplifies getting the process of getting a column
dtype.

This function maps the columns labels to its corresponding
data type. For matrices it gets the data types of aggregate
columns (aggregateColumnInfo) and for other reports, the
column types (detailColumnInfo).

In this way, I refactored two other functions. One to get a
single column data type and another to get the every column
dtype in a list.