Filter report by sorting if column is sortable
This is done to prevent any kind of limit onto the length of the filter
value, as the concatenation of previously seen values will naturally
grown at a rate of 2000 per iteration.

We will sort the identifier column if it makes sense sorting, which
here means if it is included in the list `sortable` (e.g., int,
id), and then get the next values by getting only those values
greater than the last identier column value.

For example, if the id_column is CaseId, then we will sort it in
ascending order, get all the 2000 first values. Then we get the last
value, say, 1000, and add a filter to get only CaseId's greater than
1000.

This is done is a sloppy way right now but it will be refactored.