Format filter values properly. Closes #5.
The main change here is the method Metadata.format_value, which converts dates into
ISO 8601 format if the column type is date-like or joins them together
by a comma, surrounding its values with quotes.

This prevents values with a comma conflicting with the comma used
to separate filters values, e.g. "7,01,8,02", when we really mean
'"7,01","8,02"'.

I have tested if different data types would have trouble with this.But
apparently there isn't.

The function also handles a list/tuple/pd.Series passed by the user, so
it will join its values by a comma as well. If a string is passed, it is
not treated as an iterable, being included in the filter as is.