Get full report by sorting and then filtering
This is another workaround the 2000 row limit, less vulnerable to other
limitations such as the 20 maximum filters and the maximum filter
character length.

This is how it works: if the `id_column` is a lookup field, meaning it
is some sort of a row-identifier (but it doesn't mean it is unique to
each column...), then we will first filter this column in ascending
order, get the report, and starting filtering out for the values greater
than the last one we got.